Forward port button/keyboard backlight options
Notes by @xX-Yoshi-Xx:
* First of all, the original commit is here: https://github.com/CyanogenMod/android_packages_apps_Settings/commit/6d0624b2b28ef9ed879e9a61c8946c25f1284094
* Moved the code from CM's ButtonSettings.java to our HardwareKeysSettings.java
* Cleaned up the hardware keys category code. The first category in the hardware keys
settings XML wasn't defined properly in the HardwareKeysSettings.java file. Properly define
the key for the category in the XML as a category in the java. Then, edit the button
backlight removal code to remove the button backlight option not from the root of the XML
but instead from the first category in the XML.
* Cleanup

Change-Id: Ica02fad15392201b719144c4b911407d5b370cee
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>